refactor: restructure bookList widget according to FSD
